♻️  zb: Split client handshake code into its own mod
